Add some basic methods to PyArrayDescr
