+++ tests/array.rs
@@ -247,4 +247,5 @@ fn dtype_from_py() {
         .downcast()
         .unwrap();
     assert_eq!(&format!("{:?}", dtype), "dtype('uint32')");
+    assert_eq!(dtype.get_datatype().unwrap(), numpy::DataType::Uint32);
 }
